% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popplay_code.R
\name{PopularPlays}
\alias{PopularPlays}
\title{Last 20 Songs Played by Popularity}
\usage{
PopularPlays(clientID = "", clientSecret = "")
}
\arguments{
\item{ClientID}{Given from Spotify}

\item{ClientSecret}{Given from Spotify}
}
\value{
Bar graph of popularity breakdown of last songs played
}
\description{
Function pulls data from an inputted Spotify Access Token and returns a
bar graph breaking down the popularity of each song,
100 being the most popular, 0 being the least.
}
\details{
The function shows a breakdown of the popularity of the last 20 songs
played by the user. 100 being the most popular, 0 being the least popular.
}
\examples{
PopularPlays("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX",
"XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")

}
\author{
Alexis Santise
}
\keyword{internal}
